Split Pydantic contracts by domain for #42
