+++ backend/app/contracts/__init__.py
@@ -1,0 +1,163 @@
+"""Domain-owned Pydantic contract modules."""
+
+from .requirements import *
+from .grounding import *
+from .auth import *
+from .test_cases import *
+from .execution import *
+from .impact import *
+from .projects import *
+from .orchestrator import *
+from .integrations import *
+from .exports import *
+from .automation import *
+from .reporting import *
+from .billing import *
+
+__all__ = [
+    "Requirement",
+    "RequirementExtractionItem",
+    "RequirementsOutput",
+    "BusinessRule",
+    "FieldConstraint",
+    "RolePermission",
+    "StateTransition",
+    "RiskSignal",
+    "RequirementAnalysis",
+    "ParseResponse",
+    "ReviewResult",
+    "WorkflowSettings",
+    "WorkflowDiagnostics",
+    "WorkflowIteration",
+    "ScenarioIntent",
+    "RequirementCoveragePlan",
+    "RequirementAnalysisOutput",
+    "RequirementCoveragePlanOutput",
+    "RequirementsWorkflowResponse",
+    "ArtifactSource",
+    "GroundedUIElement",
+    "GroundedApiSurface",
+    "GroundedWorkflow",
+    "GroundedContext",
+    "EnrichInput",
+    "EnrichResponse",
+    "AuthUser",
+    "GoogleLoginRequest",
+    "AuthTokenResponse",
+    "LogoutResponse",
+    "TestStep",
+    "TestCase",
+    "TestCasesOutput",
+    "TestCaseTemplate",
+    "GenerateTestCasesInput",
+    "RefineTestCasesInput",
+    "GenerateTestCasesResponse",
+    "ExecutionIssue",
+    "ExecutionUnsupportedStep",
+    "ExecutionCandidate",
+    "ExecutionPreviewInput",
+    "ExecutionPreviewSummary",
+    "ExecutionPreviewResponse",
+    "ExecutionRunInput",
+    "ExecutionRunItem",
+    "ExecutionRunSummary",
+    "ExecutionRunResponse",
+    "ImpactItemKind",
+    "ImpactChangeType",
+    "ImpactSource",
+    "ImpactRecommendationAction",
+    "ImpactChangedItem",
+    "ImpactImpactedTestCase",
+    "ImpactRecommendation",
+    "ImpactAnalysisSummary",
+    "ImpactAnalysisResult",
+    "ImpactAnalysisInput",
+    "ImpactUpdateApplyInput",
+    "ImpactUpdateApplyResult",
+    "ProjectStageName",
+    "QaProjectStageState",
+    "QaProjectSummary",
+    "QaProjectStageSnapshot",
+    "QaProjectTimelineEvent",
+    "QaProjectExecutionRun",
+    "QaProjectDetail",
+    "QaProjectListResponse",
+    "QaProjectCreateInput",
+    "QaProjectUpdateInput",
+    "QaProjectUseCaseSnapshotInput",
+    "OrchestratorStageName",
+    "OrchestratorStageStatus",
+    "OrchestratorActionId",
+    "OrchestratorBlockerCode",
+    "OrchestratorBlocker",
+    "OrchestratorActionRecommendation",
+    "OrchestratorStageState",
+    "OrchestratorStatusResponse",
+    "OrchestratorRunStatus",
+    "OrchestratorRunEventType",
+    "OrchestratorCheckpointRecord",
+    "OrchestratorRunEvent",
+    "OrchestratorRunRecord",
+    "OrchestratorRunListResponse",
+    "JiraConnectionInput",
+    "JiraStoredConnection",
+    "JiraConnectionSummary",
+    "JiraConnectionStatusResponse",
+    "JiraConnectionDeleteResponse",
+    "JiraProjectSummary",
+    "JiraProjectsResponse",
+    "JiraIssueTypeSummary",
+    "JiraProjectIssueTypesResponse",
+    "JiraIssueSummary",
+    "JiraIssueSearchResponse",
+    "JiraImportInput",
+    "JiraSyncPreviewInput",
+    "JiraSyncIssuePreview",
+    "JiraSyncPreviewResponse",
+    "JiraSyncApplyInput",
+    "JiraSyncIssueResult",
+    "JiraSyncApplyResponse",
+    "AzureDevOpsConnectionInput",
+    "AzureDevOpsStoredConnection",
+    "AzureDevOpsConnectionSummary",
+    "AzureDevOpsConnectionStatusResponse",
+    "AzureDevOpsConnectionDeleteResponse",
+    "AzureDevOpsProjectSummary",
+    "AzureDevOpsProjectsResponse",
+    "AzureDevOpsWorkItemTypeSummary",
+    "AzureDevOpsProjectWorkItemTypesResponse",
+    "AzureDevOpsWorkItemSummary",
+    "AzureDevOpsWorkItemSearchResponse",
+    "AzureDevOpsImportInput",
+    "AzureDevOpsSyncPreviewInput",
+    "AzureDevOpsSyncWorkItemPreview",
+    "AzureDevOpsSyncPreviewResponse",
+    "AzureDevOpsSyncApplyInput",
+    "AzureDevOpsSyncWorkItemResult",
+    "AzureDevOpsSyncApplyResponse",
+    "JiraExportInput",
+    "JiraExportResponse",
+    "ExportTestCasesInput",
+    "AutomationInput",
+    "AutomationResponse",
+    "UsageReportUserSummary",
+    "UsageReportGroup",
+    "UsageReportResponse",
+    "BillingCatalogEntry",
+    "BillingUserProfile",
+    "BillingAccount",
+    "BillingLedgerEntry",
+    "BillingConsumptionRecord",
+    "BillingAllocation",
+    "BillingAllocationSummary",
+    "BillingQuotaSummary",
+    "BillingWalletSummary",
+    "BillingEntitlementResponse",
+    "BillingErrorDetail",
+    "BillingCreditGrantRequest",
+    "BillingCreditGrantResponse",
+    "BillingAllocationRequest",
+    "BillingAllocationResponse",
+    "BillingLedgerResponse",
+    "BillingOrgSummaryResponse",
+]
