+++ docs/codebase/CONCERNS.md
@@ -7,7 +7,7 @@ git/history checks. It is not a full bug backlog.
 
 | Severity | Concern | Evidence | Impact | Suggested action |
 |----------|---------|----------|--------|------------------|
-| High | Large, high-churn orchestration and contract files remain central to the app | `frontend/src/App.jsx`, `backend/app/agents/test_case_agent.py`, `backend/app/models.py`, recent git history | Higher regression and merge risk for UI workflow, prompt orchestration, and API contracts | Continue behavior-preserving extraction under issue-scoped frontend, agent, and contract refactor tasks |
+| High | Large, high-churn orchestration files remain central to the app | `frontend/src/App.jsx`, `backend/app/agents/test_case_agent.py`, recent git history | Higher regression and merge risk for UI workflow and prompt orchestration | Continue behavior-preserving extraction under issue-scoped frontend and agent refactor tasks |
 | High | PostgreSQL persistence adapter, schema, and migration plan are not implemented | `docs/persistence-target-decision.md`, `backend/app/services/firestore_repository.py`, `backend/app/services/billing_repository.py`, `backend/app/services/audit_repository.py` | Audit, audit dead letters, billing, and reporting now have repository seams but still use Firestore as the transitional runtime store | Define PostgreSQL schema/migrations and add a PostgreSQL adapter behind the #49 boundaries |
 | Medium | Auth compatibility mode must stay out of production deployments | `docs/production-auth-policy-decision.md`, `backend/app/auth/jwt_auth.py`, `backend/app/routers/auth.py`, `scripts/deploy_cloud_run.sh` | Manual deployments using `firebase-or-backend-jwt` would re-enable backend JWT compatibility outside local/E2E workflows | Keep Cloud Run on `AUTH_TOKEN_MODE=firebase-only` and treat compatibility mode as local/test only |
 | Medium | Metrics endpoint exposure must stay intentionally scoped | `backend/app/main.py`, `backend/app/observability/metrics.py`, `scripts/deploy_cloud_run.sh` | `/metrics` exposes operational metadata and should not be accidentally public in production | Keep production deployments on `METRICS_ENABLED=false` or require `METRICS_ACCESS_TOKEN` plus an appropriate network perimeter |
@@ -19,7 +19,7 @@ git/history checks. It is not a full bug backlog.
 | Monolithic frontend orchestration | `App.jsx` still owns most workflow state and API actions after component extraction | `frontend/src/App.jsx` | Harder UI changes and fragile E2E updates | Extract workflow state into hooks/reducers in small issue-scoped slices |
 | Shared frontend style cascade | Feature styles are split into owned modules but still share one imported cascade | `frontend/src/styles/index.css`, `frontend/src/styles/*.css` | Cross-feature selectors and import-order changes can still cause visual regressions | Keep new selectors in the closest feature style file and cover visual workflow changes with browser checks |
 | Test-case agent orchestration | Coverage metrics, review scoring, fallback generation, and hydration are split, but prompt builders and workflow orchestration still share one module | `backend/app/agents/test_case_agent.py`, `backend/app/agents/test_case_coverage.py`, `backend/app/agents/test_case_review.py`, `backend/app/agents/test_case_fallback.py`, `backend/app/agents/test_case_hydration.py` | Prompt or workflow changes still require careful regression checks across generation behavior | Keep future prompt/orchestration changes focused and rely on helper-level tests plus offline benchmarks |
-| Broad model module | Many product domains share one Pydantic file | `backend/app/models.py` | Contract merge conflicts and long review cycles | Split model modules only after router/service ownership boundaries are stable |
+| Contract facade compatibility | Pydantic contracts now live in domain modules, but existing callers still import through the legacy facade | `backend/app/contracts/`, `backend/app/models.py` | New contract changes could drift if maintainers bypass the closest domain module or forget schema checks | Add new contracts to the relevant domain module and keep running OpenAPI export plus generated frontend API checks |
 | Partial generated frontend API types | High-traffic workflow contracts are generated from OpenAPI, but broader integrations still consume some response shapes manually | `frontend/src/api/generated/api-contracts.d.ts`, `frontend/src/App.jsx`, `scripts/generate_frontend_api_types.py` | Untyped lower-traffic calls can still drift until runtime/E2E | Extend generated contract coverage as additional API areas change |
 
 ## 3) Security Concerns
@@ -49,7 +49,7 @@ git/history checks. It is not a full bug backlog.
 | `frontend/src/App.jsx` | Owns auth, workflow state, API actions, derived metrics, and layout orchestration | Highest recent git churn among tracked source hotspots; 3013 lines | Keep changes focused and cover with frontend build/E2E |
 | `frontend/src/styles/` | Feature-owned stylesheet modules imported through a shared cascade | Former `App.css` styles now span feature files and still depend on import order | Keep selector moves scoped and verify primary workflow screens visually |
 | `backend/app/agents/test_case_agent.py` and focused test-case helper modules | Central generation orchestration plus split coverage, review, fallback, and hydration helpers | High churn; orchestrator reduced to about 1400 lines with helpers in `test_case_*.py` modules | Add focused tests for every behavior slice before changing prompt or workflow logic |
-| `backend/app/models.py` | Shared API contracts for many domains | High churn and 1047 lines | Treat changes as contract changes; run OpenAPI export |
+| `backend/app/contracts/` and `backend/app/models.py` facade | Shared API contracts are split by domain but still define the public API surface | Contract changes affect API consumers | Treat changes as contract changes; run OpenAPI export and generated frontend API checks |
 | Integration sync routers/services | JIRA and Azure DevOps have parallel flows with provider-specific edge cases | Recent traceability and test isolation work | Change one provider at a time or keep explicit parity tests |
 | Execution runtime | Converts natural language cases into runnable browser specs | New next-version E2E capability | Run execution preview and runtime checks for every change |
 
@@ -95,4 +95,5 @@ git/history checks. It is not a full bug backlog.
 - `backend/app/agents/test_case_review.py`
 - `backend/app/agents/test_case_fallback.py`
 - `backend/app/agents/test_case_hydration.py`
+- `backend/app/contracts/`
 - `backend/app/models.py`
